Remove persistent-keepalive check
Also allow for peers with an endpoint and no keepalive check to be reresolved.

The persistent keepalive is not intended for the usecase of dynamic IPs with an open endpoint as it unnecessarily increases the "chattiness" of the protocol.
Also reduce the maximum latest handshake value to 135 seconds as suggested in the upstream code from wireguard-reresolve-dns script.